feat: allow overriding of internal functions
The interesting bits of this module are the various ways of building
DAGs for files and folders (flat, sharded, etc).

Sometimes we want to utilise bits of this logic to build DAGs without
having to reimplement big chunks of this module.

This PR allows the user to pass in functions to replace key parts of
this import pipeline.

Enables:

* https://github.com/ipfs/js-ipfs-mfs/pull/73
* https://github.com/ipfs/js-ipfs-unixfs-importer/pull/46