fix: only output unixfs things
This is a `unixfs-importer` but it's possible to output CIDs that
resolve to `dag-raw` nodes if your data is small and you set
`rawLeaves` and `reduceSingleLeafToSelf` to true. In that case
you'll get a `dag-raw` node as the output.

This makes it impossible to set metadata on small files with
`rawLeaves` set to true.

BREAKING CHANGE:

If your data is below the chunk size, and you have `rawLeaves` and
`reduceSingleLeafToSelf` set to true, you'll get a CID that resolves
to a bona fide UnixFS file back with metadata and all that good
stuff instead of a `dag-raw` node.